Refactors some style things into their own package
Replaces the inline template with a `.tmpl` file and folder for the
same, allowing us a path to add new template files and partials as
necessary without continuing to make the tui file any larger.

Also refactors the styles into their own package. While this may
initially seem unnecessary and add a bit of clutter to the codebase
because now you have more `style.Error` instead of `errorStyle` - I
believe this will eventually lead to a "cleaner" and more maintainable
codebase.