Remap quit to ctrl+q to prevent accidental exit (#35)
* Fix bug silencing selected incident

Maintains the ability to send an incident list to SilenceIncidentsMsg, but adds SilenceSelectedIncidentMsg to handle individual incident silence via keypress, fixing a bug where the parsing of the selected incident didn't happen on time, causing a panic.

* remap quit to ctrl+q to prevent accidental exit

This PR remaps the quitting the app from `q` to `ctrl+q` to prevent accidental exit of the application.

Follow up work for more complicated key mapping (eg vim or tmux-style) will be done via #22

Fixes #32