Ensure incident ID is retrieved for each selected row per update
This adds a call to get the incident id of the incident
represented in the selcted row each time the table view is updated, and
this ID is used to select the incident with getSelectedIncident().

This ensures that the currently highlighted row is always the incident
that is acted upon in table view, rather than a previously selected
incident.

When no row is selected, attempts to act on an incident return a status
message to that effect, rather than panicing.

Fixes #44
Fixes #52

Signed-off-by: Chris Collins <collins.christopher@gmail.com>