Adds initial framework for tests in commands.go
This adds initial tests and a pattern to use as a framework for adding
unit tests for the rest of the `tea.Cmd` functions in `commands.go`, as well
as explanation of the pattern for future contributors.

The `tui.go` `Update()` command will receive `tea.Msg`s and return `tea.Cmd`s.
All the `tea.Cmds` (except those in the actual Bubble Tea framework) will
be collected in `commands.go`.  Each command will have an outgoing `tea.Msg`
that contains any relevant results. Commands may also have incoming
`tea.Msg`s, but these are deprecated as the `Update()` function is
converted to call commands directly rather than `func() tea.Msg {}`
commands, which are an unnecessary intermediary.

Signed-off-by: Chris Collins <collins.christopher@gmail.com>