Re-escalate to policy instead of user for silences
Changes silencing of alerts to send to an escalaton policy instead
of assigning to a specific user.

Adds framework for future change of re-escalation to an escalation.

Replaces the "silent user" and hard coded "silent escalation policy" for
silencing alerts with a configuration map.  The configuration map key
"service_escalation_policies" should consist of two REQUIRED policies,
"DEFAULT" with the policy ID for the standard policy for all alerts,
and "SILENT_DEFAULT" with the policy ID to assign assign alerts that should
be silenced.

Optionally, other PagerDuty service IDs can be provided with associated
policy IDs, to route alerts to policies based on the impacted service.

Eg:

```
---
service_escalation_policies:
  DEFAULT: P012345
  SILENT_DEFAULT: P543210
  PXYZAB: P98765
```

Signed-off-by: Chris Collins <collins.christopher@gmail.com>